Moved abstract classes to their own folder
