+++ src/abstract-api-classes/api.model.ts
